Added Debug Filter and changed Shader Uniforms
Added Debug Filter: LogFilter.
if any filter is used (so the array is not empty) only the LogTypes in the filter will be actually printed

Seperated Projection and View uniforms, for whatever reason this also fixed a rendering bug :/

Replaced forwards horizontal movement in playercontroller with directional forwards movement.

TODO
- Fix Lit shader if not already fixed
- Add Physics
- Improve Render Performance
- Add Ligth Component